Update requirements to latest version

+++ floortrans/loaders/augmentations.py
@@ -255,12 +255,12 @@ class RandomCropToSizeTorch(object):
         new_w = self.width + img_w
         new_h = self.height + img_h
 
-        new_image = torch.full([image.shape[0], new_h, new_w], 255)
+        new_image = torch.full([image.shape[0], new_h, new_w], 255, dtype=self.dtype)
         new_image[:, pad_h:img_h + pad_h, pad_w:img_w + pad_w] = image
 
-        new_rooms = torch.full((1, new_h, new_w), self.fill[0])
+        new_rooms = torch.full((1, new_h, new_w), self.fill[0], dtype=self.dtype)
         new_rooms[:, pad_h:img_h + pad_h, pad_w:img_w + pad_w] = label[0]
-        new_icons = torch.full((1, new_h, new_w), self.fill[1])
+        new_icons = torch.full((1, new_h, new_w), self.fill[1], dtype=self.dtype)
         new_icons[:, pad_h:img_h + pad_h, pad_w:img_w + pad_w] = label[1]
 
         label = torch.cat((new_rooms, new_icons), 0)
